trying to get modal to work
